requirements.txt
aiogram==3.26.0
beautifulsoup4==4.12.3
tortoise-orm==0.20.0
python-dotenv==1.0.0
gigachat==0.2.0